mediawiki::php: add daemon restart cronjob
It's not a good idea to let a php-fpm daemon run with an exhausted
opcache. While we solve the problem more radically (that means, starting
to deploy MediaWiki differently), we're left with the problem that
application servers will progressively exhaust their free opcache with
deployments. We add a daily systemd timer that:

* Checks if the free opcache is below a certain threshold
* safely restarts php-fpm

Add this class to the canaries first.

Bug: T224857
Change-Id: I6144d6b860b10dc8a471f02e6dc9c80467608d28