profile::analytics::refinery::job::data_purge: avoid saturday
Avoid saturday to drop mediawiki druid snapshots, since it is currently
a delicate operation that might cause troubles for Druid (and also
AQS as consequence). This change allows the drop script to run only
if it is a work day. The time has also been moved to 9 AM UTC, where
there are higher chances to find people around working.

Bug: T226035
Change-Id: I64a962d922705a46b9e6f5a43a94137b3ae5a327