cache: double appservers and api connection limit
For reasons not yet fully understood, under pathological conditions we
do hit the max_connections limit on appservers (1000), while we do not
in the case of restbase, which has max_connections set to 5000.

Experiment setting max_connections to 2000 for appservers and api, and
observe if it is an improvement.

Hosts: cp1075.eqiad.wmnet

Change-Id: I4b9939cc607c8d22df13202c78d8eaa77e9b21e5