cache_upload codfw: read ats-be etcd keys
Having converted cp2002 to ATS in 3b27a32, we can now make all frontends
in codfw read the etcd keys for 'ats-be'. That was not possible before
the conversion of the first codfw upload node given that confd fails if
there is no data under a given etcd key.

Hosts: cp2002.codfw.wmnet, cp2005.codfw.wmnet

Bug: T226637
Change-Id: I4ebc21ce2b22c2df42d1c3cadf05aa8c82a06d6c