varnish: temporarily allow more response headers
This is a temporary mitigation until we figure out the underlying
problem with the excessive session cookies.  We don't even know
how bad the worst cases of this are, but surely doubling the
current limit should reduce the incidence of the problem for now.

It will cost an extra ~1280 bytes per session in varnish memory
usage, which seems reasonable all things considered.

Will also push this live via varnishadm for the running instances,
this commit just makes it persist through future restarts.

Bug: T226840
Change-Id: I8f0d44ae92591ac47f8336ab5f8fa2767ea799b3