toolforge: k8s: more robust swap deletion operations
More robust operations to delete the swap are required, otherwise systemd may
act too smart and re-activate it.

There are 2 systemd mechanisms involved:
 * systemd-fstab-auto-generator
 * systemd-gpt-auto-generator

The first one automatically reads /etc/fstab and activates any swap it finds:

● dev-disk-by\x2duuid-1df5d19f\x2d38d6\x2d45e0\x2d9a07\x2d17d0e8d81672.swap [..]
   Loaded: loaded (/etc/fstab; generated; vendor preset: enabled)
   Active: active since Tue 2019-07-02 11:47:37 UTC; 1min 5s ago
     What: /dev/disk/by-uuid/1df5d19f-38d6-45e0-9a07-17d0e8d81672
     Docs: man:fstab(5)
           man:systemd-fstab-generator(8)
  Process: 293 ExecActivate=/sbin/swapon /dev/disk/by-uuid/1df5d19f-38d6-45e[..]

To prevent this, we could simply drop any swap lines from /etc/fstab.

The second one scans the disk and does the same:

● dev-vda2.swap - Swap Partition
   Loaded: loaded (/run/systemd/generator.late/dev-vda2.swap; generated; ven[..]
   Active: active since Tue 2019-07-02 11:47:37 UTC; 1min 5s ago
     What: /dev/vda2
     Docs: man:systemd-gpt-auto-generator(8)
    Tasks: 0 (limit: 4915)
   Memory: 0B
      CPU: 0
   CGroup: /system.slice/dev-vda2.swap

For this one, we need to drop the swap partition from the disk. In CloudVPS
this is usually /dev/vda2.

Bug: T215531
Change-Id: I543df30e3344d3c85803ce7ae53099475aec7a73
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>