toolforge: add a join configuration to the init setup
The join file really should be wherever the init file is.

I've removed the encryption key field for now because we were setting it
to nothing, which seems bad to me.  That field is for encrypting the certs
in a k8s secret during bootstrap of the control plane, which is probably
a good thing, but I haven't confirmed we want to specify it now.

Bug: T215531
Change-Id: I2067173e6f3f7fa309dcecd3bad0c6b592830fdc