toolforge: a firewall default of DENY is maddening for kubernetes
Within a Kubernetes cluster, all nodes should basically be able to communicate
on all ports in order to run pods everywhere.  Even control plane nodes
implement some types of pods. Micromanagement of the cluster is the exact
opposite of the scaling philosophy of kubernetes, and we have security
groups able to manage the firewall via group membership (which is fully dynamic).

I propose we manage such things entirely in SGs because the ferm classes
are a bit restrictive when you figure we have to restart ferm every time
we change puppet to add a node (also requiring lists of all the nodes etc.)

Bug: T215531
Change-Id: I47abc6c5d2247573bc2d091f383a6c1733c66c10