profile::hadoop::master: add Yarn unhealthy workers check
This change replaces a per-host status check with a single
prometheus/metric based one on the Hadoop master node.
In this way we have less alarms set, and also we don't need
a special authentication set of permissions for nagios to
check the status of every yarn node manager when kerberos is
enabled.

Bug: T226698
Change-Id: Ia4d875dad87ef81a381f3434bb4973bd99810d4d