toolforge: k8s: fix half-done code factorization for kubadm_join
In a previous patch we factorized the code, but this block which is meant to be
shared was left behind (thus, duplicated). Fix this now by simply dropping it.

Users of toolforge::k8s::kubeadm_init now have to use the class
toolforge::k8s::kubeadm_init as well. That is the case of the user
profile::toolforge::k8s::kubeadm::master.

Bug: T215531
Change-Id: Idc26caf89269c17fea77f558bea48bc012ab2614
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>