cache: reimage cp1078 as upload_ats
Use the whole nvme disk (nvme0n1) as a storage device for ATS instead of
the partition nvme0n1p1. We know that using the partition made ATS
choose 4k as the sector size on cp1076, which is what we want. This is
to confirm that ATS makes the same choice using the whole disk too, in
which case there is no need to bother with partitions at all and we can
use the whole disk everywhere.

Bug: T226638
Change-Id: Ib7baa31c3d656bd0b582d02e5c07aca2f74ce17f