toolforge: set up cert distribution for additional control plane nodes
This obviously requires that the certs for the initial node be added to
the local puppetmaster as secrets.  That seems to be a sensible way to
join new nodes. This only includes the safe ones. kubeadm should generate
the certs, using these, for the new apiserver modules.

This is triggers via hiera because we have to have some manual way of
saying the certs are there and should be used.  As long as a single control plane
node is in the cluster, we should always be able to tear down and rebuild/join
control plane nodes via this method.

Bug: T215531
Change-Id: I3b9c3da7a609e41d7afd9b02500176275abb3466