base::monitoring::host: ignore /mnt/hdfs from disk checks
The /mnt/hdfs mountpoint is a fuse read-only view of the Hadoop's
HDFS file system, that shouldn't be checked. The check triggers
a CRITICAL when Kerberos is enabled, since the mountpoint is not
accessible if the user is not authenticated.

profile::analytics::cluster::client (that is responsible to deploy
the mountpoint on analytics client nodes) already contains a check
that authenticate via Kerberos.

Bug: T226698
Change-Id: I469ebc1da84e9e31c37c557bd529404ec33f4be9