registry, swift: some images are not replicated.
As we found out in the linked task, there are some problems
about how registry uploads objects in swift and swift container synchronization.
We think that creating the same swift container in both clusters could
fix it.

This change introduces that while it does not change synchronization,
that needs to be done manually or via cumin

Bug: T227570
Change-Id: Iad66fe8680ef2b733b862f7c8640ab984037f9fa