Revert "Revert "prometheus-mysqld-exporter: Automate targets based on zarcillo db""
Because esams is outdated, and we are not happy with setting this up on bastions,
while the new profile will be present there (may be removed later after refactoring),
it will be disabled as no mysqls are setup on pops, for the reasonable future.

Original comment:
prometheus-mysqld-exporter: Automate targets based on zarcillo db

Previously, prometheus targets for prometheus-mysqld-exporter were
maintained manually.

Implement script to read the canonical place where mysql instances
are controlled (zarcillo database) and generate the files for
prometheus hosts configuration with the appropiate groups and
sections.

This reverts commit 923fbed00f934a494e1f96de64beee8f14b066f6.

Reason for revert: It will be deployed again after some fixes

Bug: T143896
Change-Id: I3081b0c934a5ffd08f1f3979e8e7d3a30a34b0f2