profile::spark2: specify the spark.ui.port and its firewall range
Port 4040 needs to be reachable by the Yarn Resource manager to fetch
the status of the Spark job (yarn.w.o -> spark job -> application master).

I tried on stat1007 to create two separate pyspark sessions, and
the ports used were 4040 and 4041. If we allow 100 separate ports
for the Spark drivers, we should allow the same for the ui.

Bug: T170826
Change-Id: Ifb19d712b559bc59c2c8402ee7625d9f40b67892