keyholder: hiera setting for require_encrypted_keys
Before the introduction of profile::keyholder::server it was possible to
override ::keyholder::require_encrypted_keys in hiera. This change
re-introduces a hiera setting that allows this to be changed on
a per-deploy basis. Allowing unencrypted keys is something that should
really only be done is testing/QA type environment, but it can be handy
there (for example in the Striker Cloud VPS project).

Change-Id: Idb46e986f832681211c1d6e16b6368d0256e25a4