Revert "Revert "Revert "Revert "prometheus-mysqld-exporter: Automate targets based on zarcillo db""""
This reverts commit 7169062bceb146d50b9c312aa62237e135ea39df.

Fixing missing '=' on template.

Original comment:

Because esams is outdated, and we are not happy with setting this up on bastions,
while the new profile will be present there (may be removed later after refactoring),
it will be disabled as no mysqls are setup on pops, for the reasonable future.

prometheus-mysqld-exporter: Automate targets based on zarcillo db

Previously, prometheus targets for prometheus-mysqld-exporter were
maintained manually.

Implement script to read the canonical place where mysql instances
are controlled (zarcillo database) and generate the files for
prometheus hosts configuration with the appropiate groups and
sections.

Reason for revert: redeploying

Bug: T143896
Change-Id: Iac388543ea2e4b22f2e37f9a2c11f4b644bd7232