hiera backends: update hiera.yaml file to work with puppet 4.9
newer version of puppet require all top level keys i.e. backends to be defined
in the backend array[1].  Currently we abuse the fact that there is a
lack of checking currently so that we can change the `datadir` based on
whether an element in the hierarchy is prefixed with `private` or
`secret`

This update fixes the config so that is compatible with newer versions
of puppet and also creates empty backend objects for secret and private.

[1]https://github.com/puppetlabs/puppet/blob/4.9.0/lib/puppet/pops/lookup/hiera_config.rb#L301-L307

Bug: T227779
Change-Id: I5b2ab74c722f107db18fb44d3ad3d91cf8a4a11b