systemd::timer::job: Add optional $max_runtime_seconds
Add a $max_runtime_seconds parameter to the systemd::timer::job define
which can be used to set a `RuntimeMaxSec=<%= @max_runtime_seconds %>`
stanza in the timer's service unit. This can be used to cap the runtime
of the job started by the timer. If the job is still running when
$max_runtime_seconds have elapsed, systemd will kill it and put the unit
in to a failure state.

This change also sets $max_runtime_seconds = 590 on the timer for
wmcs-dns-floating-ip-updater which was seen in T227785 to hang
indefinitely, thus preventing the associated timer from firing. Systemd
timers are designed not to fire if the unit they are associated with is
already running, but they do so with no signal that anything is wrong.
Since we expect to run this task every 10 minutes, we can set a 9m50s
limit on the runtime of any given invocation, and be confident that in
the worst case the task will run again in 10 minutes. We should also get
a notification of the task being killed by systemd via the monitoring
for failed timers that is already in place.

Bug: T227830
Change-Id: I7312a6130b9be6abe7727785589f78a548a4953b