Remove CherryPickCounter Diamond collector
This was initially assumed to be only used on Cloud VPS instances, but
it's actually also used on labpuppetmaster* hosts. The Diamond removal
in production is completed, so for now removing this Diamond collector
to fix Puppet runs on the labpuppetmaster* hosts. This can be re-added
as a Prometheus metric e.g. via mtail later if it's still needed.

Bug: T210993
Change-Id: Id5c31269332fbbd8d047a635797dc9f7eff128c4