jobrunners: Test php7_only on 6 jobrunners
We have been migrating async jobs to PHP7 for quite
a sometime now. This was implemented by adding special
blocks on the change-propagation/jobqueue-deploy/ repo,
that enabled the PHP7 cookie.

To migrate the rest, we will enable the php7_only on
6 eqiad jobrunners (~25%), as the blast radious will
be smaller if jobs start failing. If all goes
well, we can go ahead and migrate all jobrunners to
serve via php7 ony

Bug: T219148
Change-Id: Ifb001bab69948f3f1f53900c8b2e3af42f910812