toolforge: kubeadm master nodes shouldn't use client certs for etcd
The client cert cannot be the puppet cert anyway.  It has to be shared
by all nodes in the control plane.  The simple solution is to remove the
values, which works great!  We already have client cert verification off
on the etcd cluster, so the cert was not actually used anyway.

This is mostly a partial revert of previous commit:
 7896626f26b82a84215d13917deb45c1dab32e27

Bug: T215531
Change-Id: Icd33e6a871b113400d581c243cff843fb6248717