Introduce module amd_rocm
This module should contain all the configs to set up a AMD
GPU on a target host, refactoring it out from profile::statistics::gpu.

Since we'll want in the future to easily test new versions of ROCm
without impacting all the hosts running a GPU, the version parameter
is added.

Bug: T224723
Change-Id: Iafa732926e01d19453acfca8d280ece5b45216b8