tendril: use ldap-ro, use Hiera, refactor to profile
The reason to touch this is to stop using the "ldap-labs"
names and use the readonly server names instead.

While doing this also get the LDAP config from Hiera instead
of hardcoding it in the class.

To be able to add the parameter and not anger the style checks
we must also move parts of the role to a profile and include that.

Finally, use the new passwords::ldap::production class instead
of the deprected passwords::ldap::wmf_cluster class. You
can see in the private repo that they have the same values
but there is a TODO comment to remove the wmf_cluster one.

Bug: T227650
Change-Id: I732118fba7e84c5ee26b449e27e583a9e8f7cdf3