Add an anycast endpoint to syslog centralservers
And basic Icinga ping check monitoring.

Some appliances such as network devices or PDUs can only send syslog to
udp endpoint and not the regular pipeline.
The current setup relies on 2 endpoints: syslog.codfw.wmnet and
syslog.eqiad.wmnet. The issue if the devices are configured with only 1
endpoint (eg. the geographically closer) it is a SPOF. If the devices
are configured with both (not always supported), they the data ends up
duplicated.
Using anycast, if an endpoint goes down, logs will automatically be
routed to any other one.

Explicitely don't configure the VIP on lithium as it's going to be
decommissioned.

Matching DNS CR: https://gerrit.wikimedia.org/r/c/524045/

Todo once merged:
 - Configure neighbor IPs + static routes on the routers
 - Write documentation

Change-Id: Ib2c6c2b9a84aeb20544ec55e8e55caa84e83c792