toolforge: Enable pod security policy
This is the minimum viable code needed to spin up the cluster with psp
in place.  It exempts kube-system namespace from most restrictions so that
everything can run fine.

This tests well in my local buster cluster and should work fine.

I've added it to the kubeadm_init because there is no way it is not required
absolutely for init, and it requires no params.

Bug: T227290
Change-Id: If7b93cf38830f0135112af42299503ec9f2074f2