toolforge: Enable pod security policy
This is the minimum viable code needed to spin up the cluster with psp
in place.  It exempts kube-system namespace from most restrictions so that
everything can run fine.

This tests well in my local buster cluster and should work fine.

I've added it to the kubeadm_init because there is no way it is not required
absolutely for init, and it requires no params.

Bug: T227290
Change-Id: Ic3e432a91a3bc8b8f24c75a78e81c696dd4fd5dc