toolforge: k8s: calico_workaround: convert from profile to module
I can't find a reason for this to be a profile. This should be part of the
basic kubeadm deployment step/configuration for calico.
Also, is not reading any hiera data. So convert to a module.

This should help with ordering/dependencies, since we require the kubeadm
class to be applied before this one.

Change-Id: I60241e0e547c04b500c0b266e37f2870376bbe06
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>