Switch profile::openldap::management to obtain the LDAP server from Hiera
In 67e1e84735ba636dbd89a5cd385900e600aac7d2 eqiad cloud VMs were switched
to use the read-only replica, which had to be reverted in
0a17a9903fa1f55bc264556ad7dbfb8deb5d19f2 as user changes with the tools
provided on the mwmaint hosts broke.

This fixes the config to read the proper read/write LDAP server from Hiera.
It's an interim fix, I want to get rid of using profile::ldap::client::labs
on production hosts entirely, but that requires some deeper changes and
this one will unbreak the switch of cloud VMs to the replicas.

Bug: T46722
Change-Id: I9978b8bfbd4dc08109dda6876ceba985089632e2