puppetmaster: Add the ability to have canary backends
The CR add's the ability to configure a specific backend as a canary
puppetmaster and also provide a list of canary_hosts which should be
directed towards this host.  The initial idea for this is to test the
puppetmaster 5.5 upgrades however I think this would also be useful for
testing update and modification to the puppetmaster configuration.

The following is an example output of web-frontend.conf
 https://phabricator.wikimedia.org/P8775

Hosts: re:^puppetmaster[12]001.*\.wmnet$

Change-Id: Ibf555c88e62fec0dc4fc7d2c88db36f550f959fc