toolforge: remove class redeclaration
Because it was being forgotten in the master class, but it is essential
for kubeadm to function correctly without warnings in a supported fashion,
I made the docker service class a simple include at the module level.

Because it is not a profile, I think it is an antipattern to split it out
like this, when this profile is ultimately very simple.  The calico
workaround I think is good to keep separate because it will likely be
unnecessary in the future sometime.

I didn't expect it to be declared again at the node level, so I have to
remove it here for it to not error out.

Bug: T215531
Change-Id: I0de83e31e2702578c6b1d35fa96304808c37b7b2