toolschecker: check for existing webservice
There can only be one webservice per backend and tool running. This commit adds
a check to determine if a webservice instance is already running, and if found
wait for a short time for it to clean up.

This race condition is seen when both nagios servers overlap each other when
checking the same webservice backend.

Bug: T221301
Change-Id: Ief828e2872f453580417a083b3d71120a1ac52da