profile::tlsproxy::service: add more granularity in monitoring
In the Analytics use case (the only one that uses this profile
at the moment) we created a single TLS certificate for multiple
domains. This means that the check_https command may fire for hosts
that have nothing to do with the domain indicated by the icinga
critical message. For example, due to a feature missing in the nginx
package for buster, an alarm fired for 'yarn.wikimedia.org' for the
host (analytics-tool1004) that runs superset.wikimedia.org (listed
as SAN of yarn).

To make things more clear, I have split the check command into two
pieces:
- URI to check
- service name to check
Adding sane defaults to avoid hiera pollution for the regular use
cases.

Bug: T227860
Change-Id: I05dd3a652ed36fa9d78b722afea86e609daac6b1