Add net-tools to standard packages
In jessie, net-tools is priority:important and was automatically pulled
in. In stretch/buster that has been changed (it's now priority:optional),
but with facter 2 this was papered over by the fact that facter declared
a dependency on net-tools, so it was universally installed.

Since then we've migrated to Facter 3 which parses network settings
natively, so on recent fresh installations using stretch/buster, 
net-tools (and thus netstat, ifconfig, route, arp and a few more)
is missing.

Strictly speaking all the functionality is provided by iproute2 (ip,
ss, routel etc), but there's a number of references in our puppet repo
which refer to netstat or ifconfig (and probably some hidden ones
executed by scripts/programs without declaring a dependency).

Plus, there's muscle memory and given that net-tools is tiny, let's
just continue to keep it installed fleet-wide for now.

Change-Id: I5daf82bc61f0e11dca84538609f0200bcd00f5b6