openstack: fullstack: use a readable-friendly name for VMs
When inspecting leaks by the fullstack testing mechanisms, I find it a bit
annoying to translate the epoch to the actual date, to know when a given
VM/DNS record was created.

Solve this by use an actual date as the name, something like:

>>> int(datetime.today().strftime('%Y%m%d%H%M%S'))
20190724150140

Which can be visually translated to (at least) a concrete day.

This patch is not tested, i.e, I ignore if there are other mechanisms that
count on VMs having a concrete epoch in the name.

Change-Id: I44b35e2c80d719c2b187aca7ca2b232d806ec579
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>