profile::kerberos::kdcserver: add +requires_preauth to new users
It seems that without using +requires_preauth we get into failed
attempts to communicate with the HDFS namenode due to the absence
of PREAUTH (that we should enforce by default).

The KDC conf lists "+preauth" as default flag for all the principals,
so enforcing +requires_preauth on the kadmind side seems sound.

Bug: T226104
Change-Id: Ib11aa3b3de159c4594de7a93581b3795c755cd98