toolforge: set kubeadm to use internal registry for pause container
In order for a kubeadm-configured kubelet to use an internal registry for
the all-important pause container, it must be specified in the init.

Since we restrict users to images from the internal registry, they'll be able
to make exactly nothing work if the pause container isn't coming from it.
The existing pause containers we have up are deprecated, and there are
bugfixes and compatibilities to consider, so I've pushed the latest version (3.1)
to the toolforge registry so we can be as restrictive as needed.

This should configure kubelet for a new cluster as intended.

Bug: T228887
Bug: T215531
Change-Id: I6e10faa025d994b27722916a73a1ce76c91ee2c4