toolforge: add internal pause container to all the other kubelets
Kubelet extra args do not propogate via kubeadm join.  This is very annoying,
but the devs are not looking to change this for now. This file is sourced
by the extra file created by kubeadm for systemd as the semi-supported workaround.
The alternative is to use a join config which may or may not support it.

Bug: T228887
Bug: T215531
Change-Id: I085a8bf386d4b90c0186511a9a5c7059dc78357e