envoyproxy: create module, add tls terminator definition
We chose to create the following resources:
* A base class envoyproxy that installs the server but doesn't configure
any web resource for it
* A build-envoy-config script that is able to collect yaml fragments and
  compose them in a valid envoy configuration, and is also able to
  verify its correctness.
* An enovoyproxy::tls_terminator define to set up tls termination to
  local resources on a server
* Two envoyproxy::{listener,cluster} defines that allow to create static
  listeners and clusters

In this first iteration, only static resources are generated. It might
be viable, in the future, to also auto-generate some parts of the
definitions or to use file-based RDS.

The generated configurations might be useful also as a starting point
for sidecars in our k8s deployments for TLS termination.

Also fix the behaviour of taskgen around tox in modules - using
Dir.chdir while executing tasks in parallel can have unintended consequences.

Change-Id: Id4e58e1a4a17b5a2a74f588d20e289eadd33ea2d