Remove informational default-kubernetes-policy.yaml
This file was in the puppet repo specifically to document that it exists
and to allow git greps to work on it. Since now it is under the
deployment-charts/ repo and is actively being used in deploys, the only
use left is git grep, which can be achieved via some user education.
Remove the file

Change-Id: I3844eb8127e73e6a8a3613c1dfa22e2916f29cd3