maintain-kubeusers: set timeout on reading LDAP data
Since the script can handle LDAP not being there just fine (it fails,
which is good), but it hangs sometimes on LDAP disturbances, it seems that
we need to set a timeout on reading the data from the server.

The default timeout is "None", which would explain the behavior we see.

Bug: T194859
Change-Id: I74cbccadb85b9c306b6e403649eb52a93298eff5