maintain-kubeusers: allow ldap3 to raise exceptions
The default behavior is the reply with False, which could cause issues
in places.  Since we know we cannot set a decent timeout, and I'm
unsure if they do any threading tricks or we could set a timeout decorator,
this might help.  An exception will trigger a restart of the service
instead of just sitting there, if one actually happens.

This option is confirmed present in this version of LDAP3 python library:
https://github.com/cannatag/ldap3/blob/2aee1896cc6a884e7ae141bb45cb0a9374bec8dc/ldap3/core/connection.py#L170

Bug: T194859
Change-Id: I3603f9e114e71cb83665412d899fb1d4ff98614b