Introduce cdh::mysql_jdbc
This new define should be a good/DRY solution to deploy
the right JDBC connector package on various Debian versions.

The libmysql-java package is present up to Stretch, but not
in Buster, that has libmariadb-java.

The difference is the following (from dpkg -L):

libmysql-java:
/usr/share/java/mysql-connector-java.jar
/usr/share/java/mysql.jar

libmariadb-java:
/usr/share/java/mysql-connector-java-5.1.42.jar

As far as I can see the new jar should work for both.

Change-Id: I445a92e27ca3eeb7a0eb351fbf7d664d48e51862