toolforge: k8s: add nginx-ingress configuration.
This patch contains a basic configuration to add PodSecurityPolicy role binding
for the nginx-ingress service account.
It also contains all the initial nginx-ingress configuration in the form of a
concatenated yaml file build from the upstream docs.

The TLS config is included. However we won't do TLS termination at
nginx-ingress in this first iteration of the setup. If the TLS secret is not
included nginx-ingress will refuse to start.

Bug: T228500
Bug: T227290
Change-Id: Ic49f4e707e2f0aaa8b3f9a23e854793f6a87a998
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>