anycast recdns: edge sites via realm.pp (nop)
This is a followup improvement on the hieradata work in 80ad9a0a ,
and should be a nop.  It's clearly cleaner (for now) to just use
the existing realm.pp switch statements to set per-site defaults while
we continue the transition here, instead of messing with the complex
precedence rules of hieradata while preserving the special-case entries
for the recdnses themselves.

Hopefully at a later date, we can improve the recdns puppetization
to the point that they can use the anycast themselves as well (which
means we'd need the loopback anycast IP's existence to be predicated on
a running and healthy pdns recursor daemon), at which point we can
probably remove this hieradata setting completely, at least for
production purposes.

Bug: T228190
Change-Id: If2a5f5fcd91999306c93aed9ef2453515631cd7e