mediawiki: use a better notes_url for the "DSH groups" Icinga alert
As described on the ticket the current URL is very outdated and does
not give correct information on how to handle the alerts
that say a Mediawiki appserver is "not in mediawiki-installation
DSH groups".

It still talks about the actual DSH group files that were once in the repo.

Nowadays we use conftool though and this is just about conftool-data and
being pooled or not.

Made this new page for it.

Bug: T227547
Change-Id: I282cd3c7b16bb91091ec05219ae1a236e19d889f