vcl: add Access-Control-Allow-Origin to mobile redirects
Clients bound by the same-origin policy (e. g. in-browser XHR or
fetch()) should be allowed to read this redirect and attempt to follow
it – whether they’re actually allowed to access the target is then
determined at the time of the second, redirected request.

One case where we need this is accessing Wikidata’s Special:EntityData
from another wiki – Special:EntityData always makes its response public,
but if the request is made from a mobile client, Varnish redirects it to
the mobile subdomain first, and the client needs to be allowed to read
that redirect.

Bug: T229385
Change-Id: I11f2a8fd5ff9e286f1209e5188d6cc180c79f491