hiera backends: update the config and hiera backend with the correct names
This change attempts to re-deploy a failed change[1] updating the hiera backends
again.  When the initial change was deployed a number of systems started to alert in
icinga with 'Failed to apply catalog, zero resources tracked by'.  I believe this was
caused as i did not deploy the change to the puppet masters simultaneously first.
This results in a node potentially hitting two puppet masters with different
configurations.

I believe the initial issue was transient and would have resolved it self.  however a
misunderstanding in how `run-puppet-agent --failed-only` operated led me to believe
nodes where continuing to fail spasmodically.  I have now updated run-puppet-agent[2]
to also consider a catalogue with zero resources as a failure.

I will follow the following procedure when re-implementing
  1. disable puppet fleet wide
  2. merge change
  3. perform an agent run on the puppet masters to deploy change
  4. restart apache on puppet masters
  5. test change on a few canary hosts
  6. enable puppet fleet wide

 [1]I340591bb43f9ec0a4c64d5f1167c3834fd36a2aa
 [2]I2e87af7ab76a01133552a93922cdd9300a8b7bae

Change-Id: I9003bbfbb68068698b4fbe0c7ce8b76dfe6908a3