Update wmflabs.org redirect target
wmflabs.org currently points to an on-wiki redirect - when visited,
the end page displays "(Redirected from Portal:Wikimedia Labs)",
which is confusing because that wasn't the original page visited.
However, discussion on the patch concluded that, instead of just
bypassing the redirect, `Help:Cloud_Services_Introduction` would
be a better target for the site's hard redirect.

Bug: T229896
Change-Id: I38bed589eba17170cd369985767d87fe9d7b4684