Remove obsolete/unsupported nobarrier option from three partman recipes
In Linux 4.19 nobarrier got removed from XFS for which Filippo pushed
a patch to adapt the fstab generated on Swift backends running Buster.

We also have three partman recipes passing the option, but it turns
out this isn't actually passed down to the installed hosts:

The filesystem gets created by partman-xfs and partman only supports
a whitelisted range of file system options, visible at
https://salsa.debian.org/installer-team/partman-xfs/blob/master/mountoptions/xfs

nobarrier wasn't ever supported there, so remove it to reduce confusion
(and it would fail if ever passed to a Buster host as well).

Bug: T229915
Change-Id: I92eed34a41f5a2f251b2eaab1712cf626b6c51bd