puppetmaster::frontend: update web conf to use RewriteRules instead of proxypass
As we need to use mod_rewrite to proxy the canary_hosts we need to
convert everything to RewriteRules.  This is because mod_rewrite hooks
before mod_proxy meaning that canary hosts will be sent to the canary
server for everything.

In this CR I have also removed the rules for the old api endpoints.
As far as I can tell only /puppet/v3 and puppet-ca/v1/certificate are
required.

Note: if anyone knows how to do a conditional ProxyPass for the canary_hosts
i would prefer to go that route but i was unable to come up with a solution
that worked

Bug: T228657
Change-Id: Ia7570fb54ba3e473358f29729d56c746b01d8743