anycast recdns: config for 41 eqiad canaries
Most clusters don't have a designated canary.  While the anycast
recdns changes seem very low-risk so far, we'd like to take the
intermediate step of hitting limited machines from a broader set
of various cluster configurations before turning on either of the
core DCs as a whole.

This is a manually-curated list built from looking at site.pp and
extracting 1/N of various eqiad clusters that seem like important
or interesting validation cases, and trying to avoid "master"
nodes where possible in asymmetric clusters.

The many per-host hieradatas created in this commit will be
reverted in the final patch which turns on anycast recdns for all
of eqiad, completing the global turn-up.

Bug: T228190
Change-Id: Ic02959e0fb05bb665490a581c57d0f67e2564a50