profile::tlsproxy::envoy: new TLS terminator for services
As an alternative to tlsproxy::localssl which is nowadays used both as a
proxy for internal services and for external traffic at the edge, we
start using Envoy, a modern, robust proxy optimized for
service-to-service interaction. In this first installment, it only
supports TLS termination to local services. Proxying to external
services will be added at a later stage.

Also add directly the configuration for the envoy jobs to prometheus, in
order to collect the copious telemetry it exposes.

Bug: T210411
Change-Id: I7eedc2ddc782bacb679fe0288cb253bbbbf89fbe