monitoring: change WMCS services to paging WMCS only
This also includes removing the plane alert for 1m load numbers to 5m.
Changing to prometheus without recording rules to build subqueries around
that like what was done in graphite alerting is just not good.  This is
slightly better.  The rest of this should just be removing critical and
selecting our team.

Bug: T229884
Change-Id: I81a9a6bfdb871bb135f54453cf191eb6d4b142e7