puppetmaster upgrade: add a lua filter to remove the job_id
The new puppet masters add a job_id to all reports.  however the
puppetdb4.4 schema dose not recognise this.  this hack adds a lua filter to
nginx to remove the job_id from the payload before sending it to
puppetdb

I have created a simple test running at http://18.221.94.108/test/ which
runs this lua code with a simple backend flask app which simply does
  `return request.data`

examples of the failed reports can be found in:
 puppetdb1001.eqiad.wmnet:/root/tmp

if i remove the `"job_id":null,` from one of theses files.  then move it
to `/var/lib/puppetdb/stockpile/cmd/q/` and finally restart puppetdb.
then the report seems to be processed correctly so there is a chance
this could work but it is, in every way, a hack!!!

ping me on irc if you want a login to 18.221.94.108

Bug: T230002
Change-Id: Iadf8eda334c3b90606e0de7bd368ba713b26f3da