labstore: restore original sense of the load alert with prometheus
Learning from Jason and experimenting, this should operate roughly the
same way as the old metric we got from graphite to monitor 1m load while
still smoothing spikes in a way that is quieter than a simple average
would be.  We want only things that are high 85% of the time period, not
just a high overall average.

Bug: T229884
Change-Id: Icbff96f1f4ae9571d630836fff9897813d91ad71