systemd::timer::job: fix the default logging paths
For compatibility reasons with the originating analytics specific
define, the log_basedir parameter was let set to a value that didn't
make sense. In fact, the timer "foo" would've logged in
"/var/log/foo/foo/foo.log", but only if "/var/log/foo" was previously
created.

As a consequence, with default parameters one could expect a puppet
failure as the parent directory wasn't found.

Fixing this to a saner "/var/log/foo/syslog.log", which will anyways
have no effect as most timers have been declared with logging disabled,
and the ones that use it override base_logdir anyways.

Change-Id: I7f161471f4ecff2fd622db6a5bd9944f43e0fa8c