profile:templates:services_proxy: Enable ipv6 and listen only locally
We had noticed that as were serving more traffic
via PHP7, the more socket errors on a server were increasing.
Specifically, we had a rate of about 2 failed connection attempts
per second.

After some investigation, we found out that php7 was trying
to connect to the local elasticsearch proxy (in this case, nginx
listing on ipv4 only) on ::1. Since nginx was listening to
0.0.0.0, this was a failed connection attempt.

This patch enables ipv6 on services_proxy. Additionally, we are
limiting services_proxy to listen only to local interfaces as that's
where we should be accepting connections.

After merging, this requires restarting nginx across the fleet.

Bug: T224538
Change-Id: I07799f19c44772124d0322a04a5138234dc2fc92