Add Cache-Control response header for Wikistats V2's index.html
The wikistats V2 website is a single page (index.html) website.
During a deployment its content may change, for example changing
a CSS or a JS file, therefore its content should not be cached
if possible for long periods of time.

This change forces index.html to be cached (even by browsers)
for a maximum of 10s before asking for a revalidation
(asking to the upstream server if the resource has been modified
since, leveraging the Etag response header for example).

Bug: T230136
Change-Id: I62c5e1a86fa7fe4e93a48d1c751606b14319a433