ATS: enable compress plugin on cp5002
As an experiment, re-enable compress.so on cp5002 to see if introducing
minimum-content-length has any effect on the crashes.

If not, we will try to set 'cache' to false in compress.config to ensure
that ATS only caches the compressed or decompressed variant returned by
the origin, and not both as alternates.

Bug: T227432
Change-Id: I061f480e898d29051a2161cfe23bdf6c4ed6e0f8