scap: set flag for check-and-restart-php
Now that we have scap 3.12.0-1 deployed, we can set the feature flag to
trigger php7 opcache restart checks at the end of each deployment. Only
set for production at the moment, would need to add the restart script
to beta before we were able to add it to "[wmflabs]" or "[global]".

Bug: T224857
Change-Id: I8ea44b0662f6361a41aa3db654e5f7788e067dbb