envoyproxy: only run build-envoy-config from puppet
I realized it's not a great idea to have our configuration files
writable from the daemon that needs to read them. Thus, remove the
verify-and-build step from the envoy service unit, and moved it to only
happen in puppet. Also, change the owner of all config files back to root.

Change-Id: Ia99de69f7bebf521b9b65c777077f9d885ccd0a7