profile::analytics::cluster::packages::common: temp remove python3-tk
In T229347 the Analytics team is experimenting with the deployment of
python3.7 on Stretch nodes to ease the migration to Spark2. The main problem
is that the version of python3-tk on stretch requires python3.5, and puppet
keeps removing/adding packages because of that. Temporary removing this
package to avoid excessive churning in puppet runs.

Bug: T229347
Change-Id: Ia133b0f3cbcf487f66b2fced41a6854e72f7e0b4