ATS: enable compress.so for upload@eqsin
cp5002 has been running with the compress plugin enabled for the past 3
days without crashes. It seems that the segfaults where due to caching
of both the compressed and decompressed variant. At the very least,
disabling the feature with a30d645 made the issue go away.

Now enable compress.so for all upload@eqsin.

Bug: T227432
Change-Id: I58a9e57452bd1ab316e143cb23c2f4e33c56b663