cdh::hive: re-allow localjoins in hiveserver2's jvm
This change removes the global hive.auto.convert.join = false
setting, re-enabling it since the default is true.
To prevent issues with the current Prometheus configuration
of the Hive 2 server, hive.exec.submit.local.task.via.child
is explicitly set to false to force the execution of the local
optimizations in the same jvm.

Bug: T209536
Change-Id: Ide45aea60c296d387a81f4be3ac34abff94d2641